feat: implement published-week

+++ backend/src/usecases/shiftUsecase.ts
@@ -1,4 +1,5 @@
 import * as shiftRepository from "../database/default/repository/shiftRepository";
+import * as publishedWeekRepository from "../database/default/repository/publishedWeekRepository";
 import { Between, FindManyOptions, FindOneOptions, LessThanOrEqual, MoreThanOrEqual, Not } from "typeorm";
 import Shift from "../database/default/entity/shift";
 import { ICreateShift, IFindShift, IUpdateShift } from "../shared/interfaces";
@@ -31,8 +32,10 @@ export const create = async (payload: ICreateShift): Promise<Shift> => {
     endTime: shift.endTime,
     startTime: shift.startTime,
   });
-
   if (isExist) throw new Error("Shift is overlapping with another shift");
+
+  const isPublished = await publishedWeekRepository.findByDate(shift.date);
+  if (isPublished) throw new Error("Couldn't create shift, the week has been published")
 
   return shiftRepository.create(shift);
 };
@@ -47,8 +50,10 @@ export const updateById = async (
     startTime: payload.startTime,
     id,
   });
-
   if (isExist) throw new Error("Shift is overlapping with another shift");
+
+  const isPublished = await publishedWeekRepository.findByDate(payload.date);
+  if (isPublished) throw new Error("Couldn't update shift, the week has been published")
 
   return shiftRepository.updateById(id, {
     ...payload,
@@ -56,5 +61,11 @@ export const updateById = async (
 };
 
 export const deleteById = async (id: string | string[]) => {
+  const _id = Array.isArray(id) ? id[0] : id;
+  const shiftData = await shiftRepository.findById(_id);
+
+  const isPublished = await publishedWeekRepository.findByDate(shiftData.date);
+  if (isPublished) throw new Error("Couldn't delete shift, the week has been published")
+
   return shiftRepository.deleteById(id);
 };
